+++ src/Decoder.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace SandFox\Bencode;
 
+use SandFox\Bencode\Exceptions\FileNotReadableException;
+
 final class Decoder
 {
     private array $options;
@@ -97,10 +99,14 @@ final class Decoder
      */
     public function load(string $filename): mixed
     {
+        if (!is_file($filename) || !is_readable($filename)) {
+            throw new FileNotReadableException('File does not exist or is not readable: ' . $filename);
+        }
+
         $stream = fopen($filename, 'r');
 
         if ($stream === false) {
-            return false;
+            throw new FileNotReadableException('Error reading file: ' . $filename); // @codeCoverageIgnore
         }
 
         $decoded = self::decodeStream($stream);