+++ CHANGELOG.md
@@ -87,6 +87,13 @@
 
 ## 1.x
 
+### 1.7.3
+
+*Oct 24, 2021*
+
+* dump() now throws exception if the file is not writable
+* load() now throws exception if the file is not readable
+
 ### 1.7.2
 
 *Oct 23, 2021*